google_agent_development_kit/mock_weather_agent/pyproject.toml
[project]
name = "weather-bot"
version = "0.1.0"
description = "A unified weather bot supporting both ADK CLI and custom Python services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.9.0",
    "litellm>=1.74.15.post1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "flask>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]

[project.scripts]
weather-bot = "src.weather_bot.services.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weather_bot"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
